+++ NEWS.md
@@ -9,6 +9,10 @@ from the US EPA, AirNow, AIRSIS, WRCC and others.
 
 ## Version 1.0 -- Monitoring Data Access
 
+### PWFSLSmoke 1.0.21
+
+ * removed errant TAB character from AQI_es$names
+
 ### PWFSLSmoke 1.0.20
 
  * tweaks for CRAN submission